chore: Refactor Books endpoint for CRUD operations.
