+++ backend/app/crud/__init__.py
@@ -1,3 +1,4 @@
 """ CRUD package """
 # Import all modules
 from . import user
+from . import book
